+++ backend/src/main/java/com/apex/backend/dto/PaperAccountResetRequest.java
@@ -3,6 +3,8 @@ package com.apex.backend.dto;
 import lombok.Data;
 
 @Data
+import java.math.BigDecimal;
+
 public class PaperAccountResetRequest {
-    private Double startingCapital;
+    private BigDecimal startingCapital;
 }
