+++ backend/src/main/java/com/apex/backend/dto/PaperStatsDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -14,8 +15,8 @@ public class PaperStatsDTO {
     private Integer winningTrades;
     private Integer losingTrades;
     private Double winRate;
-    private Double totalProfit;
-    private Double totalLoss;
-    private Double netPnl; // ✅ FIXED: Changed netPnL to netPnl to match Controller
-    private Double profitFactor; // ✅ Added to prevent potential future error
+    private BigDecimal totalProfit;
+    private BigDecimal totalLoss;
+    private BigDecimal netPnl; // ✅ FIXED: Changed netPnL to netPnl to match Controller
+    private BigDecimal profitFactor; // ✅ Added to prevent potential future error
 }
